requirements.txt
numpy==2.1.3
pillow==11.0.0
pygame==2.6.1
pyserial==3.5